Web client(Dimasik)/webclient/requirements.txt
django==4.2.16
django-redis==5.2.0
gunicorn==20.1.0
requests==2.31.0
